Reorganize layouts and content
Page layouts had redundant code such as CSS and JS tags in the head, though with page-specific tags mixed in. Common tags are now in the base layout, and most pages use that layout. Their content is now included in either their markdown or HTML page with front matter.

Pages with lots of text have been converted to markdown when possible for source readability and editability.