Update layouts and download pages
Page layouts have been moved into the pages themselves. Download pages for v1 and v2 have been filled out, with links at the top of the download pages to the other download pages.

A link to the Rat TWAS Hub is added to the navbar.

Fixes #7 since the site is ready to go live with v3.